+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Update Custom Pod Autoscaler version to `v2.2.0`.
+- Update Go version to `v1.16`.
 
 ## [v0.6.0] - 2020-08-31
 ### Changed